Solution for exercise 4.21

+++ part_4/blog_list/controllers/blogs.js
@@ -21,7 +21,7 @@ blogsRouter.post('/', async (request, response) => {
 
 	const decodedToken = jwt.verify(request.token, process.env.SECRET)
   if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
+    return response.status(401).json({ error: 'Token invalid' })
   }
   const user = await User.findById(decodedToken.id)
 
@@ -44,8 +44,23 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+	const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'Token invalid' })
+  }
+  const user = await User.findById(decodedToken.id)
+
+  const blogToDelete = await Blog.findById(request.params.id)
+	if (!blogToDelete) {
+		return response.status(404).json({ error: 'Blog not found' })
+	}
+	if (blogToDelete.user.toString() === user.id.toString()) {
+		await Blog.findByIdAndDelete(request.params.id)
+		response.status(204).end()
+	}
+	else {
+		response.status(403).json({ error: 'Forbidden: You have no rights to delete this blog' })
+	}
 })
